Moved login and logout methods to correct paths
